refactor: cvc number 에러 수정 및 리팩토링
