+++ src/components/CardCvcNumber/CardCvcNumber.tsx
@@ -1,12 +1,11 @@
-import InputText from "../InputText/InputText";
+import { useState } from "react";
 import NumberInput from "../Input/CardNumberInput";
 import InputErrorMessage from "../Input/InputErrorMessage";
-import useCvcNumberInputHandler from "../../hooks/useCvcNumber/useCvcNumberInputHandler";
+import InputText from "../InputText/InputText";
 
 interface CardCvcNumberProps {
-  handleChange: (value: string) => void;
-  cvcNumbers: string;
-  errorMessage: string;
+  numbers: string;
+  setNumbers: React.Dispatch<React.SetStateAction<string>>;
   onComplete: () => void;
 }
 
@@ -16,17 +15,47 @@ const CARD_CVC_NUMBER = {
   PLACEHOLDER: "123",
 } as const;
 
+const CVC_RULE = {
+  INVALID_LENGTH_ERROR: "CVC 번호는 3자리 숫자여야 합니다.",
+  MAX_LENGTH: 3,
+} as const;
+
 export default function CardCvcNumber({
-  handleChange,
-  cvcNumbers,
-  errorMessage,
+  numbers,
+  setNumbers,
   onComplete,
 }: CardCvcNumberProps) {
-  const { handleCardCvcNumberChange } = useCvcNumberInputHandler(
-    errorMessage,
-    handleChange,
-    onComplete
-  );
+  const [error, setError] = useState("");
+
+  const cvcNumbersValidate = (value: string): string => {
+    if (value.length > CVC_RULE.MAX_LENGTH) {
+      return error; // 무시하고 기존 error 유지
+    }
+
+    if (value.length < CVC_RULE.MAX_LENGTH) {
+      return CVC_RULE.INVALID_LENGTH_ERROR;
+    }
+
+    return ""; // 정확히 3자리면 에러 없음
+  };
+
+  const handleCardCvcNumberChange = (value: string) => {
+    if (value.length > CVC_RULE.MAX_LENGTH) {
+      return; // 4자리 넘어가면 입력 아예 무시
+    }
+
+    const validationError = cvcNumbersValidate(value);
+
+    setNumbers(value);
+    setError(validationError);
+
+    const isCompleteLength = value.length === CVC_RULE.MAX_LENGTH;
+    const isNoError = validationError === "";
+
+    if (isCompleteLength && isNoError) {
+      onComplete();
+    }
+  };
 
   return (
     <section>
@@ -35,11 +64,11 @@ export default function CardCvcNumber({
       <NumberInput
         onChange={handleCardCvcNumberChange}
         placeholder={CARD_CVC_NUMBER.PLACEHOLDER}
-        value={cvcNumbers}
-        errorMessage={errorMessage}
+        value={numbers}
+        errorMessage={error}
         type="text"
       />
-      <InputErrorMessage message={errorMessage} />
+      <InputErrorMessage message={error} />
     </section>
   );
 }
